moved halo discovery into halos folder
