+++ include/cstone/halos/discovery.hpp
