+++ test/unit/halos/discovery.cpp
@@ -31,7 +31,7 @@
 
 #include "gtest/gtest.h"
 
-#include "cstone/domain/halodiscovery.hpp"
+#include "cstone/halos/discovery.hpp"
 #include "cstone/tree/octree_util.hpp"
 
 using namespace cstone;
